Extracted all useConstraints code to separate files
