+++ src/useConstraints/useConstraints.ts
@@ -1,198 +1,15 @@
 import { useMemo } from 'react';
-import { MutableValues, ConstantValues } from '../useGenericValues';
+import { ConstantValues } from '../useGenericValues';
 import {
   Validator, Validators, SingleValidator, validateValidators,
 } from '../useValidation';
-import { assert, LoggingTypes } from '../utils';
 import { Errors } from '../useErrors';
 import { FieldValue, Fields } from '../useFields';
 import {
-  supportedConstraints,
-  RequiredConstraint,
-  supportedTypesSet,
-  supportedProperties,
-  defaultMessage,
   ConstraintValidators,
   SyncedConstraint,
-  Constraints,
-  constraintValues,
-  supportedTypes,
 } from './types';
-
-function hasValue(value?: FieldValue): value is string {
-  return value === 0 || Boolean(value);
-}
-
-function getRuleValueAndMessage(
-  rules: Constraints, name: supportedConstraints,
-): [constraintValues, string] {
-  const rule = rules[name];
-  let message = '';
-  let value: constraintValues;
-  if (Array.isArray(rule)) {
-    ([value, message] = rule);
-  } else {
-    value = rule || '';
-  }
-  if (name === 'required') {
-    message = !message && typeof value === 'string' ? value : message;
-    value = Boolean(value);
-  }
-  return [value, message];
-}
-
-function getRuleValue(rules: Constraints, name: supportedConstraints): constraintValues {
-  const [value] = getRuleValueAndMessage(rules, name);
-  return value;
-}
-
-const propertyValidators = {
-  required(_: Constraints, required: constraintValues, value?: FieldValue): boolean {
-    return !required || hasValue(value);
-  },
-  maxLength(_: Constraints, maxLength: constraintValues, value?: FieldValue): boolean {
-    return !hasValue(value) || (typeof value === 'string' && value.length <= Number(maxLength));
-  },
-  minLength(_: Constraints, minLength: constraintValues, value?: FieldValue): boolean {
-    return !hasValue(value) || (typeof value === 'string' && value.length >= Number(minLength));
-  },
-  max(rules: Constraints, max: constraintValues, value?: FieldValue): boolean {
-    if (!hasValue(value)) {
-      return true;
-    }
-    const type = getRuleValue(rules, 'type');
-    return (type === 'date')
-      ? new Date(value) <= new Date(max as string | number)
-      : Number(value) <= Number(max);
-  },
-  min(rules: Constraints, min: constraintValues, value?: FieldValue): boolean {
-    if (!hasValue(value)) {
-      return true;
-    }
-    const type = getRuleValue(rules, 'type');
-    return (type === 'date')
-      ? new Date(value) >= new Date(min as string | number)
-      : Number(value) >= Number(min);
-  },
-  type(_: Constraints, type: constraintValues, value: FieldValue = ''): boolean {
-    if (!hasValue(value)) {
-      return true;
-    }
-    let regex: RegExp;
-    switch (type) {
-    case 'url':
-      regex = /^((([A-Za-z]{3,9}:(?:\/\/)?)(?:[-;:&=+$,\w]+@)?[A-Za-z0-9.-]+|(?:www\.|[-;:&=+$,\w]+@)[A-Za-z0-9.-]+)((?:\/[+~%/.\w\-_]*)?\??(?:[-+=&;%@.\w_]*)#?(?:[.!/\\\w]*))?)$/;
-      return regex.test(value);
-    case 'email':
-      regex = /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/;
-      return regex.test(value);
-    case 'number':
-      return /^-?(\d+|\d+\.\d+|\.\d+)([eE][-+]?\d+)?$/.test(value);
-    case 'date':
-      // TODO: update to date yyyy-MM-dd
-      // TODO: support datetime-local and datetime with warning
-      return !Number.isNaN((new Date(value)).getTime());
-    case 'text':
-    default: return true;
-    }
-  },
-  pattern(_: Constraints, pattern: constraintValues, value: FieldValue = ''): boolean {
-    return !hasValue(value) || (!(pattern instanceof RegExp) || pattern.test(value));
-  },
-};
-
-function getRuleMessage(rules: Constraints, name: supportedConstraints): string {
-  const [, message] = getRuleValueAndMessage(rules, name);
-  return message;
-}
-function hasRule<T extends supportedConstraints>(
-  rules: Constraints, name: T,
-): rules is RequiredConstraint<T> {
-  return ({}).hasOwnProperty.call(rules, name);
-}
-
-function validateRule(name: string, rules: Constraints): void {
-  assert.error(
-    !!rules && typeof rules === 'object',
-    LoggingTypes.typeError,
-    `(expected: Constraints, received: ${typeof rules}) The Constraints object with name (${name}) is invalid.`,
-  );
-  // throws warnings for invalid rules
-  if (hasRule(rules, 'type')) {
-    const type = getRuleValue(rules, 'type') as supportedTypes;
-    assert.warning(
-      supportedTypesSet.has(type),
-      LoggingTypes.constraintError,
-      `(input: ${name}) unsupported type (${type}). An unsupported type just means we don't have custom validation logic for this.`,
-    );
-  }
-  if (hasRule(rules, 'maxLength') && hasRule(rules, 'minLength')) {
-    const minLength = getRuleValue(rules, 'minLength');
-    const maxLength = getRuleValue(rules, 'maxLength');
-    assert.warning(
-      maxLength >= minLength,
-      LoggingTypes.constraintError,
-      `(input: ${name}) maxLength (${maxLength}) is less than minLength (${minLength}).`,
-    );
-  }
-  if (hasRule(rules, 'min') && hasRule(rules, 'max')) {
-    const min = getRuleValue(rules, 'min');
-    const max = getRuleValue(rules, 'max');
-    assert.warning(
-      max >= min,
-      LoggingTypes.constraintError,
-      `(input: ${name}) max (${max}) is less than min (${min}).`,
-    );
-  }
-  if (hasRule(rules, 'pattern')) {
-    const pattern = getRuleValue(rules, 'pattern');
-    assert.warning(
-      pattern instanceof RegExp,
-      LoggingTypes.constraintError,
-      `(input: ${name}) pattern must be a RegExp object.`,
-    );
-  }
-}
-function validateUsingContraints(rules: Constraints, value?: FieldValue): string {
-  // check required
-  const erroredProperty = supportedProperties.find((property): boolean => {
-    let hasError = false;
-    if (hasRule(rules, property)) {
-      const propValue = getRuleValue(rules, property);
-      hasError = !propertyValidators[property](rules, propValue, value);
-    }
-    return hasError;
-  });
-  let message = '';
-  if (erroredProperty) {
-    message = getRuleMessage(rules, erroredProperty) || defaultMessage[erroredProperty];
-  }
-  return message;
-}
-
-function mapConstraintsToValidators(rules: ConstraintValidators): Validators {
-  assert.error(
-    rules && typeof rules === 'object',
-    LoggingTypes.invalidArgument,
-    `(expected: Object<string, Constraint> received: ${typeof rules}) ${mapConstraintsToValidators.name} requires a constraint object.`,
-  );
-  return Object.keys(rules).reduce((
-    validationMap: MutableValues<Validator>,
-    name: string,
-  ): MutableValues<Validator> => {
-    const currentValidator = rules[name];
-    if (typeof currentValidator !== 'function') {
-      validateRule(name, currentValidator);
-    }
-    // eslint-disable-next-line no-param-reassign
-    validationMap[name] = (typeof currentValidator !== 'function')
-      ? (value?: FieldValue): string => (
-        validateUsingContraints(currentValidator, value)
-      )
-      : currentValidator;
-    return validationMap;
-  }, {});
-}
+import { mapConstraintsToValidators } from './validators';
 
 // TODO: clean up docs if necessary
